central icon server
set up a central icon server for the app to download icons from (found in /IconsServer)
to run:
-open up a terminal & cd to IconsServer
-make sure nodejs is installed, if not run `brew install node` (brew found @ https://brew.sh)
-run `npm i` and `npm i typescript --save-dev`
-run `npm run dev`
thats it!
in-app function (getIcon) contains an `endpoint` parameter which contains a URL for the server. currently, this is set to http://localhost:3000, but when we get a server setup, change `localhost` to the server IP