this is gonna make development semi-easier in the future
i just wanted to make the distinction between a system app and a user-added app. if the paywall is added & a free user would only be able to add a certain number of apps, we don't want built-in apps to be included in that. this will make paywalling easier for us in the long run